refacto: extraction de listHashesToAddAndDelete dans son propre fichier
